+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file contains the list of changes made to the Joulescope driver.
 
 ## 1.3.0
 
-2023 Mar 30 [in progress]
+2023 Mar 30
 
 * Added pyjoulescope_driver.time64 module (from UI).
 * Fixed buffer_signal summary_get handling on zero size.