+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 This file contains the list of changes made to the Joulescope driver.
 
 
+## 1.3.1
+
+2023 Apr 4 [in progress]
+
+* Decreased JS110 status polling interval to reduce USB message spamming.
+* Added JS110 streaming when only statistics requested (uses host-side stats).
+
+
 ## 1.3.0
 
 2023 Mar 30